test: テストを main_test.go に統合し個別テストファイルを削除
filter_test.go / format_test.go / service_test.go を削除し、
main_test.go に全テストを統合。git.go の parseRemoteURL と
extractSHAFromBlame のテストも新規追加（全25件 PASS）。
